+++ internal/mkdeb/mkdeb.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
+	"time"
 
 	"github.com/TecharoHQ/yeet/internal"
 	"github.com/TecharoHQ/yeet/internal/pkgmeta"
@@ -63,7 +64,13 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 			continue
 		}
 
-		contents = append(contents, &files.Content{Type: files.TypeDir, Destination: d})
+		contents = append(contents, &files.Content{
+			Type:        files.TypeDir,
+			Destination: d,
+			FileInfo: &files.ContentFileInfo{
+				MTime: time.Date(1970, time.January, 1, 0, 0, 0, 0, time.UTC),
+			},
+		})
 	}
 
 	for repoPath, osPath := range p.ConfigFiles {
@@ -72,7 +79,8 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 			Source:      repoPath,
 			Destination: osPath,
 			FileInfo: &files.ContentFileInfo{
-				Mode: os.FileMode(0600),
+				Mode:  os.FileMode(0600),
+				MTime: time.Date(1970, time.January, 1, 0, 0, 0, 0, time.UTC),
 			},
 		})
 	}
@@ -82,6 +90,9 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 			Type:        files.TypeFile,
 			Source:      repoPath,
 			Destination: filepath.Join("/usr/share/doc", p.Name, rpmPath),
+			FileInfo: &files.ContentFileInfo{
+				MTime: time.Date(1970, time.January, 1, 0, 0, 0, 0, time.UTC),
+			},
 		})
 	}
 
@@ -90,6 +101,9 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 			Type:        files.TypeFile,
 			Source:      repoPath,
 			Destination: rpmPath,
+			FileInfo: &files.ContentFileInfo{
+				MTime: time.Date(1970, time.January, 1, 0, 0, 0, 0, time.UTC),
+			},
 		})
 	}
 
@@ -102,7 +116,14 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 			return nil
 		}
 
-		contents = append(contents, &files.Content{Type: files.TypeFile, Source: path, Destination: path[len(dir)+1:]})
+		contents = append(contents, &files.Content{
+			Type:        files.TypeFile,
+			Source:      path,
+			Destination: path[len(dir)+1:],
+			FileInfo: &files.ContentFileInfo{
+				MTime: time.Date(1970, time.January, 1, 0, 0, 0, 0, time.UTC),
+			},
+		})
 
 		return nil
 	}); err != nil {